Fix -Wshorten-64-to-32 error in MatMulNBitsBiasFusion test
K and N became runtime const int64_t (router shape selection), so passing
them to MlasBlockwiseQuantizedShape/BufferSizes (int params) now triggers
-Werror,-Wshorten-64-to-32 on macOS/arm64 and Android. Cast to int.

+++ onnxruntime/test/optimizer/graph_transform_test.cc
@@ -10817,12 +10817,12 @@ TEST_F(GraphTransformationTests, MatMulNBitsBiasFusion) {
 
       int q_rows, q_cols;
       MlasBlockwiseQuantizedShape<float, qbits>(block_size, /* columnwise */ true,
-                                                K, N,
+                                                static_cast<int>(K), static_cast<int>(N),
                                                 q_rows, q_cols);
 
       size_t q_data_size_in_bytes, q_scale_size, q_zp_size_in_bytes;
       MlasBlockwiseQuantizedBufferSizes<qbits>(block_size, /* columnwise */ true,
-                                               K, N,
+                                               static_cast<int>(K), static_cast<int>(N),
                                                q_data_size_in_bytes, q_scale_size, &q_zp_size_in_bytes);
 
       auto* A = builder.MakeInput<float>(std::vector{M, K}, "A");
